docs: use shared @antdv-next/docs-plugins package
- 移除 packages/docs/plugins/markdown 本地插件,统一使用独立工程
  /Users/carl/Desktop/antdv-next/docs-plugins (link: 引用)
- markdown 层 demo 锚点使用 headerMode: examples(与 antdv-next 一致)
- TS→JS 格式化改为注入 oxfmt js-formatter
- 保留 registry 虚拟模块、源码 HTML 预渲染等 x 站点选项
