--- packages/docs/plugins/markdown/index.ts
@@ -1,8 +1,0 @@
-import type { PluginOption } from "vite-plus";
-
-import { demoPlugin } from "./demo";
-import { md2VuePlugin } from "./md2vue";
-
-export function mdPlugin(): PluginOption[] {
-  return [md2VuePlugin(), demoPlugin()];
-}
