--- packages/docs/plugins/markdown/plugins/image.ts
@@ -1,24 +1,0 @@
-import type MarkdownIt from "markdown-it";
-
-import { EXTERNAL_URL_RE } from "../shared";
-
-export interface ImagePluginOptions {
-  lazyLoading?: boolean;
-}
-
-export function imagePlugin(
-  md: MarkdownIt,
-  { lazyLoading }: ImagePluginOptions = {},
-) {
-  const imageRule = md.renderer.rules.image!;
-  md.renderer.rules.image = (tokens, idx, options, env, self) => {
-    const token = tokens[idx]!;
-    let url = token.attrGet("src");
-    if (url && !EXTERNAL_URL_RE.test(url)) {
-      if (!/^\.?\//.test(url)) url = `./${url}`;
-      token.attrSet("src", decodeURIComponent(url));
-    }
-    if (lazyLoading) token.attrSet("loading", "lazy");
-    return imageRule(tokens, idx, options, env, self);
-  };
-}
