--- packages/docs/plugins/markdown/plugins/pre-wrapper.ts
@@ -1,45 +1,0 @@
-import type MarkdownIt from "markdown-it";
-
-export interface Options {
-  hasSingleTheme: boolean;
-}
-
-export function preWrapperPlugin(md: MarkdownIt, options: Options) {
-  const fence = md.renderer.rules.fence!;
-  md.renderer.rules.fence = (...args) => {
-    const [tokens, idx] = args;
-    const token = tokens[idx]!;
-
-    token.info = token.info.replace(/\[.*\]/, "");
-
-    const active = / active(?: |$)/.test(token.info) ? " active" : "";
-    token.info = token.info.replace(/ active$/, "").replace(/ active /, " ");
-
-    const lang = extractLang(token.info);
-    const rawCode = fence(...args);
-    return `<div class="language-${lang}${getAdaptiveThemeMarker(options)}${active}"><button title="Copy Code" class="copy"></button><span class="lang">${lang}</span>${rawCode}</div>`;
-  };
-}
-
-export function getAdaptiveThemeMarker(options: Options) {
-  return options.hasSingleTheme ? "" : " ant-code-theme";
-}
-
-export function extractTitle(info: string, html = false) {
-  if (html)
-    return (
-      info.replace(/<!--[\s\S]*?-->/g, "").match(/data-title="(.*?)"/)?.[1] ||
-      ""
-    );
-  return info.match(/\[(.*)\]/)?.[1] || extractLang(info) || "txt";
-}
-
-function extractLang(info: string) {
-  return info
-    .trim()
-    .replace(/=(\d*)/, "")
-    .replace(/:(no-)?line-numbers([{ ]|$|=).*/, "")
-    .replace(/(-vue|\{| ).*$/, "")
-    .replace(/^vue-html$/, "template")
-    .replace(/^ansi$/, "");
-}
