--- packages/docs/plugins/markdown/shared.ts
@@ -1,4 +1,0 @@
-export const EXTERNAL_URL_RE = /^(?:[a-z]+:|\/\/)/i;
-
-export const SCRIPT_REGEX = /<script\b[^>]*>[\s\S]*?<\/script>/gi;
-export const STYLE_REGEX = /<style\b[^>]*>[\s\S]*?<\/style>/gi;
