+++ packages/docs/vite.config.ts
@@ -1,5 +1,6 @@
 import { AntdvNextResolver } from "@antdv-next/auto-import-resolver";
 import { AntdvNextXResolver } from "@antdv-next/auto-import-resolver-x";
+import { mdPlugin } from "@antdv-next/docs-plugins";
 import vue from "@vitejs/plugin-vue";
 import vueJsx from "@vitejs/plugin-vue-jsx";
 import { fileURLToPath, URL } from "node:url";
@@ -11,12 +12,26 @@ import { tsxResolveTypes } from "vite-plugin-tsx-resolve-types";
 import vueResolveTypes from "vite-plugin-vue-resolve-types";
 import { defineConfig } from "vite-plus";
 
-import { mdPlugin } from "./plugins/markdown";
+import { createOxfmtJsFormatter } from "./plugins/js-formatter";
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
-    mdPlugin(),
+    mdPlugin({
+      markdown: {
+        // demo 锚点统一挂到 `## Examples` / `## 代码演示 {#examples}` 标题下（与 antdv-next 约定一致）
+        demo: {
+          headerMode: "examples",
+        },
+        externalLink: true,
+      },
+      demo: {
+        virtualMode: "registry",
+        renderSourceHtml: true,
+        sourceVersionOnHmr: true,
+        jsFormatter: createOxfmtJsFormatter(),
+      },
+    }),
     vueResolveTypes(),
     vue({
       include: [/\.vue$/, /\.md$/],
